+++ volume/utils_misc.go
@@ -16,18 +16,18 @@ func getRandVolume() *Volinfo {
 	v.ID = uuid.NewRandom()
 	v.Name = fmt.Sprintf("volume-%d", atomic.AddUint64(&volCount, 1))
 	v.Type = DistReplicate
-	brickCount := uint64(rand.Intn(256) + 1)
-	for i := uint64(1); i <= brickCount; i++ {
+	brickCount := rand.Intn(256) + 1
+	for i := 1; i <= brickCount; i++ {
 		//v.Bricks = append(v.Bricks, fmt.Sprintf("host:/brick-%d", i))
 		v.Bricks[i].Hostname = "Host"
 		v.Bricks[i].Path = fmt.Sprintf("/brick-%d", i)
 		v.Bricks[i].ID = v.ID
 	}
-	v.DistCount = uint64(rand.Intn(256) + 1)
-	v.ReplicaCount = uint16(rand.Intn(10))
-	v.StripeCount = uint16(rand.Intn(10))
-	v.DisperseCount = uint16(rand.Intn(10))
-	v.RedundancyCount = uint16(rand.Intn(10))
+	v.DistCount = rand.Intn(256) + 1
+	v.ReplicaCount = rand.Intn(10)
+	v.StripeCount = rand.Intn(10)
+	v.DisperseCount = rand.Intn(10)
+	v.RedundancyCount = rand.Intn(10)
 
 	v.Status = VolCreated
